CPlot: ondrag, Intersector: clean up
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4533 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ KCore/KCore/Nuga/include/mpi_messages.hxx
@@ -659,7 +659,7 @@ namespace NUGA
       {
         STACK_ARRAY(MPI_Request, NB_REQ_FOR_HASSENT, sreqs);
         int req_count = -1;
-        for (size_t r = 0; r < nranks; ++r)
+        for (int r = 0; r < nranks; ++r)
         {
           if (r == rank) continue;
           int err = MPI_Irecv(&has_sent[r], 1, MPI_C_BOOL, r, int(TAG_HASSENT), COM, &sreqs[++req_count]);
@@ -674,7 +674,7 @@ namespace NUGA
       
       // 1. catch all sizes :
       int NB_SENT = 0;
-      for (size_t r = 0; r < nranks; ++r) if (has_sent[r])++NB_SENT;
+      for (int r = 0; r < nranks; ++r) if (has_sent[r])++NB_SENT;
 
       int NB_REQ_FOR_SIZES = 5 * NB_SENT;
       STACK_ARRAY(MPI_Request, NB_REQ_FOR_SIZES, sreqs_sz);
@@ -691,7 +691,7 @@ namespace NUGA
 
       int req_count = -1;
       int rcount = -1;
-      for (size_t r = 0; r < nranks; ++r)
+      for (int r = 0; r < nranks; ++r)
       {
         if (has_sent[r] == false) continue;
         ++rcount;
